Move exportmodifiers to dbcog
